Moved models into data folder
