+++ lib/data/repository/shuttle_repository.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:latlong/latlong.dart';
 
+import '../../data/models/shuttle_route.dart';
+import '../../data/models/shuttle_stop.dart';
+import '../../data/models/shuttle_update.dart';
 import '../../global_widgets/shuttle_arrow.dart';
-import '../../models/shuttle_route.dart';
-import '../../models/shuttle_stop.dart';
-import '../../models/shuttle_update.dart';
 import '../provider/shuttle_api_provider.dart';
 //import '../provider/shuttle_local_provider.dart';
