+++ lib/pages/routes_page/detail_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_map/flutter_map.dart';
-
 import 'package:latlong/latlong.dart';
+
 import '../../blocs/on_tap/on_tap_bloc.dart';
 import '../../blocs/theme/theme_bloc.dart';
-import '../../models/shuttle_stop.dart';
+import '../../data/models/shuttle_stop.dart';
 import 'widgets/panel.dart';
 
 class DetailPage extends StatefulWidget {
