+++ lib/pages/routes_page/widgets/panel.dart
@@ -6,7 +6,7 @@ import 'package:latlong/latlong.dart';
 
 import '../../../blocs/on_tap/on_tap_bloc.dart';
 import '../../../blocs/theme/theme_bloc.dart';
-import '../../../models/shuttle_stop.dart';
+import '../../../data/models/shuttle_stop.dart';
 import 'shuttle_line.dart';
 
 class Panel extends StatefulWidget {
